[plot] Move all the ZScale logic to independent module, then add license
- Add zscale module and move implementations
- Give full license to top of zscale file
- Add data() method to image
- Make coerce_to_array_view2 public for use in zscale module
